Excepción Tiempo de Respuesta Invalido

+++ AppListener.java
@@ -25,7 +25,7 @@ public interface AppListener {
 	void rellenarCliente(String id);
 	
 	// Ventana Editar Ticket
-	void editarTicketGUI(String idCliente, String idTicket, String nuevaDescripcion, String nuevoEstado, int nuevaSatisfaccion, int horasResolucion);
+	void editarTicketGUI(String idCliente, String idTicket, String nuevaDescripcion, String nuevoEstado, int nuevaSatisfaccion, double horasResolucion);
 	
 	// Ventana Eliminar Ticket
 	void EliminarTicket(String idCliente, String idTicket);
